remove string handling code
