+++ lib/constants.js
@@ -1,13 +1,13 @@
 
 module.exports = {
     types: {
-        object: '\00',
-        string: '\01',
-        int: '\02'
+        object: 0x00,
+        string: 0x01,
+        int: 0x02
     },
     special: {
-        objectEnd: '\b',
-        stringEnd: '\00',
-        propertyNameEnd: '\00'
+        objectEnd: 0x08,
+        stringEnd: 0x00,
+        propertyNameEnd: 0x00
     }
 };
