+++ lib/parser.js
@@ -2,33 +2,25 @@ var _ = require('lodash');
 var constants = require('./constants');
 
 
-function execRegexAt(str, reg, i){
-    reg.lastIndex = i;
-    return reg.exec(str);
-}
-
 function isNumber(str){
     return str && (str.match(/\d+/)||'').length;
 }
 
 function parseStringValue(input){
-    var m = execRegexAt(input.str, /([^\00\01\02\b]+)?\00/y, input.i);
-    var value = m[1];
-    input.i += (value ? value.length: 0) + 1;
+    var start = input.i;
+    while(input.buf.readUInt8(input.i) != constants.special.stringEnd){
+        input.i++;
+    }
+
+    var value = input.buf.toString('utf8', start, input.i);
+    input.i++;
     return value;
 }
 
 // LE 32 bit number
-// each character is one byte
 function parseIntValue(input){
 
-    var chars = [
-        input.str.charCodeAt(input.i),
-        input.str.charCodeAt(input.i + 1),
-        input.str.charCodeAt(input.i + 2),
-        input.str.charCodeAt(input.i + 3)
-    ];
-    var value = Buffer.from(chars).readInt32LE();
+    var value = input.buf.readInt32LE(input.i);
     input.i += 4;
     return value;
 }
@@ -39,17 +31,15 @@ function parseObjectValue(input){
 
     //should have properties, with type and value
     do {
-        var type = input.str[input.i];
-       
+        var type = input.buf.readUInt8(input.i);
+        input.i++;
         //handle empty objects
         if (type === constants.special.objectEnd) {
             break;
         }
 
-        var m = execRegexAt(input.str, /([^\00\01\02\b]+)\00/y, ++input.i);
-        var propName = m[1];
+        var propName = parseStringValue(input);
 
-        input.i += 1 + propName.length;
         try {
             var val;
             switch (type) {
@@ -72,12 +62,12 @@ function parseObjectValue(input){
             }
             obj[propName] = val;
         } catch (err) {
-           throw new Error('encountered error while handling property: ' + propName +' '+ err);
+            throw new Error('encountered error while handling property: ' + propName +' '+ err);
         }
 
-    } while(input.i < input.str.length);
+    } while(input.i < input.buf.length);
 
-    input.i++;
+    //input.i++;
 
     if(input.options.autoConvertArrays) {
         //check if we should convert to an array instead
@@ -96,8 +86,8 @@ function parseObjectValue(input){
 }
 
 
-function topLevel(str, opts){
-    return parseObjectValue({str:str,i: 0, options: opts || { autoConvertArrays: true, autoConvertBooleans: true }});
+function topLevel(buffer, opts){
+    return parseObjectValue({buf:buffer, i: 0, options: opts || { autoConvertArrays: true, autoConvertBooleans: true }});
 }
 
 module.exports = topLevel;
