+++ lib/writer.js
@@ -3,7 +3,15 @@ var _ = require('lodash');
 
 var constants = require('./constants');
 
-function writeValue(val){
+function stringValues(str){
+    if(!str){
+        return [];
+    }
+    var buffer = Buffer.from(str);
+    return Array.from(buffer.values());
+}
+
+function appendValue(val, octets){
 
     //pre process values to align with writable values
     if(_.isNull(val) || _.isUndefined(val)) {
@@ -19,15 +27,17 @@ function writeValue(val){
 
     // write values
     if(_.isString(val)) {
-        return val + constants.special.stringEnd;
+        var values = stringValues(val);
+        values.push(constants.special.stringEnd);
+        return octets.concat(values);
     } else if(_.isNumber(val)){
         // four bytes, LE
         var buffer = Buffer.alloc(4);
         buffer.writeInt32LE(val);
-        return buffer.toString();
+        return octets.concat(Array.from(buffer.values()));
     } else if(_.isObject(val)){
 
-        var str = '';
+        var values = [];
         var keys;
         if(_.isArray(val)){
             keys = _.range(val.length);
@@ -40,26 +50,29 @@ function writeValue(val){
             var propValue = val[key];
 
             if(_.isNull(propValue) || _.isUndefined(propValue) || _.isString(propValue)){
-                str += constants.types.string;
+                values.push(constants.types.string);
             } else if(_.isBoolean(propValue) || _.isNumber(propValue) || _.isDate(propValue)){
-                str += constants.types.int;
+                values.push(constants.types.int);
             } else {
-                str += constants.types.object;
+                values.push(constants.types.object);
             }
-            str+= key + constants.special.propertyNameEnd;
 
-            str += writeValue(propValue);
+            values = values.concat(stringValues(key.toString()));
+            values.push(constants.special.propertyNameEnd);
+
+            values = appendValue(propValue, values);
         });
-        str += constants.special.objectEnd;
+        values.push(constants.special.objectEnd);
 
-        return str;
+        return octets.concat(values);
     }
 
     throw new Error('Writer encountered unhandled value: '+val);
 }
 
 function topLevel(obj){
-    return writeValue(obj);
+    var octets = appendValue(obj, []);
+    return Buffer.from(octets);
 }
