+++ test/read-write-tests.js
@@ -14,14 +14,18 @@ function readFile(fileName, cb){
     });
 }
 
+function getFilePath(fileName){
+    return __dirname + '/static/'+fileName+'.vdf';
+}
+
 describe('read-write tests', function(){
 
     it('reads and writes same content: shortcuts1', function(done){
 
 
-        readFile('shortcuts1', function(input){
+        shortcut.parseFile(getFilePath('shortcuts1'), function(err, obj, buffer){
 
-            var obj = shortcut.parse(input, { autoConvertBooleans: true, autoConvertArrays: true });
+            should.equal(null, err);
 
             var expected = {
                 shortcuts: [
@@ -29,8 +33,8 @@ describe('read-write tests', function(){
                         AppName: 'Minecraft - FTB',
                         exe: '"U:\\Users\\Me\\Minecraft\\Feed the Beast\\launcher^FTB_Launcher.exe"',
                         StartDir: '"U:\\Users\\Me\\Minecraft\\Feed the Beast\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -40,8 +44,8 @@ describe('read-write tests', function(){
                         AppName: 'Minecraft',
                         exe: '"U:\\Users\\Me\\Minecraft\\Vanilla\\Minecraft_New.exe"',
                         StartDir: '"U:\\Users\\Me\\Minecraft\\Vanilla\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -51,8 +55,8 @@ describe('read-write tests', function(){
                         AppName: 'Tom Clancy\'s The Division',
                         exe: '"C:\\Program Files (x86)\\Ubisoft\\Ubisoft Game Launcher\\games\\Tom Clancy\'s The Division\\TheDivision.exe"',
                         StartDir: '"C:\\Program Files (x86)\\Ubisoft\\Ubisoft Game Launcher\\games\\Tom Clancy\'s The Division\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -63,9 +67,9 @@ describe('read-write tests', function(){
 
             should.deepEqual(expected, obj);
 
-            var output = shortcut.stringify(obj);
+            var output = shortcut.writeBuffer(obj);
 
-            should.equal(input, output);
+            should.deepEqual(buffer, output);
             done();
 
         });
@@ -75,9 +79,9 @@ describe('read-write tests', function(){
     it('reads and writes same content: shortcuts1 (no auto arrays)', function(done){
 
 
-        readFile('shortcuts1', function(input){
+        shortcut.parseFile(getFilePath('shortcuts1'), { autoConvertBooleans: true }, function(err, obj, buffer){
 
-            var obj = shortcut.parse(input, { autoConvertBooleans: true });
+            should.equal(null, err);
 
             var expected = {
                 shortcuts: {
@@ -85,8 +89,8 @@ describe('read-write tests', function(){
                         AppName: 'Minecraft - FTB',
                         exe: '"U:\\Users\\Me\\Minecraft\\Feed the Beast\\launcher^FTB_Launcher.exe"',
                         StartDir: '"U:\\Users\\Me\\Minecraft\\Feed the Beast\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -96,8 +100,8 @@ describe('read-write tests', function(){
                         AppName: 'Minecraft',
                         exe: '"U:\\Users\\Me\\Minecraft\\Vanilla\\Minecraft_New.exe"',
                         StartDir: '"U:\\Users\\Me\\Minecraft\\Vanilla\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -107,8 +111,8 @@ describe('read-write tests', function(){
                         AppName: 'Tom Clancy\'s The Division',
                         exe: '"C:\\Program Files (x86)\\Ubisoft\\Ubisoft Game Launcher\\games\\Tom Clancy\'s The Division\\TheDivision.exe"',
                         StartDir: '"C:\\Program Files (x86)\\Ubisoft\\Ubisoft Game Launcher\\games\\Tom Clancy\'s The Division\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -119,9 +123,9 @@ describe('read-write tests', function(){
 
             should.deepEqual(expected, obj);
 
-            var output = shortcut.stringify(obj);
+            var output = shortcut.writeBuffer(obj);
 
-            should.equal(input, output);
+            should.deepEqual(buffer, output);
             done();
 
         });
@@ -131,9 +135,9 @@ describe('read-write tests', function(){
     it('reads and writes same content: shortcuts1 (no auto booleans)', function(done){
 
 
-        readFile('shortcuts1', function(input){
+        shortcut.parseFile(getFilePath('shortcuts1'), { autoConvertArrays: true }, function(err, obj, buffer){
 
-            var obj = shortcut.parse(input, { autoConvertArrays: true });
+            should.equal(null, err);
 
             var expected = {
                 shortcuts: [
@@ -141,8 +145,8 @@ describe('read-write tests', function(){
                         AppName: 'Minecraft - FTB',
                         exe: '"U:\\Users\\Me\\Minecraft\\Feed the Beast\\launcher^FTB_Launcher.exe"',
                         StartDir: '"U:\\Users\\Me\\Minecraft\\Feed the Beast\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: 0,
                         AllowDesktopConfig: 1,
                         OpenVR: 0,
@@ -152,8 +156,8 @@ describe('read-write tests', function(){
                         AppName: 'Minecraft',
                         exe: '"U:\\Users\\Me\\Minecraft\\Vanilla\\Minecraft_New.exe"',
                         StartDir: '"U:\\Users\\Me\\Minecraft\\Vanilla\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: 0,
                         AllowDesktopConfig: 1,
                         OpenVR: 0,
@@ -163,8 +167,8 @@ describe('read-write tests', function(){
                         AppName: 'Tom Clancy\'s The Division',
                         exe: '"C:\\Program Files (x86)\\Ubisoft\\Ubisoft Game Launcher\\games\\Tom Clancy\'s The Division\\TheDivision.exe"',
                         StartDir: '"C:\\Program Files (x86)\\Ubisoft\\Ubisoft Game Launcher\\games\\Tom Clancy\'s The Division\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         IsHidden: 0,
                         AllowDesktopConfig: 1,
                         OpenVR: 0,
@@ -175,9 +179,9 @@ describe('read-write tests', function(){
 
             should.deepEqual(expected, obj);
 
-            var output = shortcut.stringify(obj);
+            var output = shortcut.writeBuffer(obj);
 
-            should.equal(input, output);
+            should.deepEqual(buffer, output);
             done();
 
         });
@@ -187,10 +191,9 @@ describe('read-write tests', function(){
     it('reads and writes same content: shortcuts2', function(done){
 
 
-        readFile('shortcuts2', function(input){
+        shortcut.parseFile(getFilePath('shortcuts2'), function(err, obj, buffer){
 
-            var obj = shortcut.parse(input);
-
+            should.equal(null, err);
 
             var expected = {
                 shortcuts: [
@@ -198,9 +201,9 @@ describe('read-write tests', function(){
                         AppName: 'AI Suite 3',
                         exe: '"C:\\Program Files (x86)\\ASUS\\AI Suite III\\AISuite3.exe"',
                         StartDir: '"C:\\Program Files (x86)\\ASUS\\AI Suite III\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
-                        LaunchOptions: undefined,
+                        icon: '',
+                        ShortcutPath: '',
+                        LaunchOptions: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -212,9 +215,9 @@ describe('read-write tests', function(){
 
             should.deepEqual(expected, obj);
 
-            var output = shortcut.stringify(obj);
+            var output = shortcut.writeBuffer(obj);
 
-            should.equal(input, output);
+            should.deepEqual(buffer, output);
             done();
 
         });
@@ -225,9 +228,9 @@ describe('read-write tests', function(){
     it('reads and writes same content: shortcuts3', function(done){
 
 
-        readFile('shortcuts3', function(input){
+        shortcut.parseFile(getFilePath('shortcuts3'), function(err, obj, buffer){
 
-            var obj = shortcut.parse(input);
+            should.equal(null, err);
 
             var expected = {
                 shortcuts: [
@@ -235,9 +238,9 @@ describe('read-write tests', function(){
                         AppName: 'AI Suite 3',
                         exe: '"C:\\Program Files (x86)\\ASUS\\AI Suite III\\AISuite3.exe"',
                         StartDir: '"C:\\Program Files (x86)\\ASUS\\AI Suite III\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
-                        LaunchOptions: undefined,
+                        icon: '',
+                        ShortcutPath: '',
+                        LaunchOptions: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -248,9 +251,9 @@ describe('read-write tests', function(){
                         AppName: '3D Vision Photo Viewer',
                         exe: '"C:\\Program Files (x86)\\NVIDIA Corporation\\3D Vision\\nvstview.exe"',
                         StartDir: '"C:\\Program Files (x86)\\NVIDIA Corporation\\3D Vision\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
-                        LaunchOptions: undefined,
+                        icon: '',
+                        ShortcutPath: '',
+                        LaunchOptions: '',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
@@ -262,9 +265,9 @@ describe('read-write tests', function(){
 
             should.deepEqual(expected, obj);
 
-            var output = shortcut.stringify(obj);
+            var output = shortcut.writeBuffer(obj);
 
-            should.equal(input, output);
+            should.deepEqual(buffer, output);
             done();
 
         });
@@ -275,10 +278,9 @@ describe('read-write tests', function(){
     it('reads and writes same content: shortcuts4', function(done){
 
 
-        readFile('shortcuts4', function(input){
+        shortcut.parseFile(getFilePath('shortcuts4'), { autoConvertBooleans: true, autoConvertArrays: true, dateProperties: ['LastPlayTime']}, function(err, obj, buffer){
 
-            console.log(JSON.stringify({str: input}));
-            var obj = shortcut.parse(input, { autoConvertBooleans: true, autoConvertArrays: true, dateProperties: ['LastPlayTime']});
+            should.equal(null, err);
 
             var expected = {
                 shortcuts: [
@@ -286,21 +288,21 @@ describe('read-write tests', function(){
                         AppName: 'The Legend of Zelda: Breath of the Wild',
                         exe: '"C:\\Games\\CEMU\\Cemu.exe"',
                         StartDir: '"C:\\Games\\CEMU\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         LaunchOptions: '-f -g "C:\\Games\\CEMU Roms\\The Legend of Zelda Breath of the Wild\\00050000101C9500\\code\\U-King.rpx"',
                         IsHidden: false,
                         AllowDesktopConfig: true,
                         OpenVR: false,
-                        LastPlayTime: new Date('2017-05-01 09:24:13.000 -0400'),
+                        LastPlayTime: new Date('2017-05-01 09:24:07.000 -0400'),
                         tags: []
                     },
                     {
                         AppName: 'Xenoblade Chronicles X',
                         exe: '"C:\\Games\\CEMU\\Cemu.exe"',
                         StartDir: '"C:\\Games\\CEMU\\"',
-                        icon: undefined,
-                        ShortcutPath: undefined,
+                        icon: '',
+                        ShortcutPath: '',
                         LaunchOptions: '-f -g "C:\\Games\\CEMU Roms\\Xenoblade Chronicles X\\code\\spaceTravel.rpx"',
                         IsHidden: false,
                         AllowDesktopConfig: true,
@@ -313,9 +315,46 @@ describe('read-write tests', function(){
 
             should.deepEqual(expected, obj);
 
-            var output = shortcut.stringify(obj);
+            var output = shortcut.writeBuffer(obj);
 
-            should.equal(input, output);
+            should.deepEqual(buffer, output);
+            done();
+
+        });
+
+    });
+
+
+    it('reads and writes same content: shortcuts5', function(done){
+
+
+        shortcut.parseFile(getFilePath('shortcuts5'), { autoConvertBooleans: true, autoConvertArrays: true, dateProperties: ['LastPlayTime']}, function(err, obj, buffer){
+
+            should.equal(null, err);
+
+            var expected = {
+                shortcuts: [
+                    {
+                        appname: '中文',
+                        exe: '"C:\\Games\\CEMU\\Cemu.exe"',
+                        StartDir: '"C:\\Games\\CEMU\\"',
+                        icon: '',
+                        ShortcutPath: '',
+                        LaunchOptions: '',
+                        IsHidden: false,
+                        AllowDesktopConfig: true,
+                        OpenVR: false,
+                        LastPlayTime: new Date('2017-05-05 15:48:09.000 -0400'),
+                        tags: [ 'Test' ]
+                    }
+                ]
+            };
+
+            should.deepEqual(expected, obj);
+
+            var output = shortcut.writeBuffer(obj);
+
+            should.deepEqual(buffer, output);
             done();
 
         });
